+++ packages/typescript/configs/index.ts
@@ -1,0 +1,23 @@
+import eslintComments from './eslint-comments.js';
+import javaScript from './javascript.js';
+import json from './json.js';
+import json5 from './json5.js';
+import n from './n.js';
+import packageJson from './package-json.js';
+import simpleImportSort from './simple-import-sort.js';
+import typeScript from './typescript.js';
+import unicorn from './unicorn.js';
+import yaml from './yaml.js';
+
+export default {
+  eslintComments,
+  javaScript,
+  json,
+  json5,
+  packageJson,
+  simpleImportSort,
+  n,
+  typeScript,
+  unicorn,
+  yaml,
+};