+++ packages/typescript/configs/simple-import-sort.ts
@@ -1,6 +1,8 @@
 import type { Linter } from 'eslint';
+import simpleImportSortPlugin from 'eslint-plugin-simple-import-sort';
 
-export const simpleImportSortRules: Linter.RulesRecord = {
+const rules: Linter.RulesRecord = {
+  'sort-imports': 'off',
   'simple-import-sort/exports': 'warn',
   'simple-import-sort/imports': ['warn', {
     groups: [
@@ -16,4 +18,11 @@ export const simpleImportSortRules: Linter.RulesRecord = {
       ['^[^.]'], // scss imports
     ],
   }],
+};
+
+export default {
+  plugins: {
+    'simple-import-sort': simpleImportSortPlugin,
+  },
+  rules,
 };