+++ packages/typescript/configs/yaml.ts
@@ -1,0 +1,29 @@
+import type { Linter } from 'eslint';
+import rawYamlPlugin from 'eslint-plugin-yml';
+import yamlParser from 'yaml-eslint-parser';
+
+import { getSafeLinterPlugin } from '../utils/isLinterPlugin.js';
+
+const yamlPlugin = getSafeLinterPlugin(rawYamlPlugin);
+
+const rules: Linter.RulesRecord = {
+  'yml/quotes': ['error', { prefer: 'single', avoidEscape: true }],
+  'yml/no-empty-document': 'off',
+  'spaced-comment': 'off',
+};
+
+const config: Linter.Config = {
+  files: ['**/*.{yaml,yml}'],
+  languageOptions: {
+    parser: yamlParser,
+  },
+  plugins: {
+    yml: yamlPlugin,
+  },
+  rules: {
+    ...yamlPlugin.configs.recommended?.rules,
+    ...rules,
+  },
+};
+
+export default config;